third_party/WebKit/LayoutTests/fast/scrolling/keyboard-scroll-page-scale-expected.txt

Test that keyboard scrolling while the page is scaled scrolls both viewports. To test manually, pinch zoom into the page and use the arrow keys, page up/down, home/end to scroll the page. You should be able to reach the end of the page bounds (i.e. scroll to see the divs at the bounds.)

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS successfullyParsed is true

TEST COMPLETE
Testing arrow keys:
PASS window.scrollY is > 0
PASS window.scrollY is pageHeight - window.innerHeight
PASS window.scrollY is 0
PASS window.scrollX is pageWidth - window.innerWidth
PASS window.scrollX is 0

Testing home and end keys:
PASS window.scrollY is pageHeight - window.innerHeight
PASS window.scrollY is 0

Testing page up and page down keys:
PASS window.scrollY is > 0
PASS window.scrollY is pageHeight - window.innerHeight
PASS window.scrollY is 0

Testing spacebar:
PASS window.scrollY is > 0
PASS window.scrollY is pageHeight - window.innerHeight
PASS window.scrollY is 0

Top of pageBottom of pageLeft of pageRight of pageMiddle of page
